+++ net/machinemuse/general/gui/clickable/ClickableModule.java
@@ -48,7 +48,7 @@ public class ClickableModule extends Clickable {
 	}
 
 	@Override
-	public boolean hitBox(int x, int y) {
+	public boolean hitBox(double x, double y) {
 		boolean hitx = Math.abs(x - getPosition().x()) < 8;
 		boolean hity = Math.abs(y - getPosition().y()) < 8;
 		return hitx && hity;
